requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
executing==1.2.0
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.2
ipdb==0.13.13
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
msgpack==1.0.5
neovim==0.3.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pynvim==0.4.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
